+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunTable.tsx
@@ -132,7 +132,7 @@ export const RunTable = (props: RunTableProps) => {
         }
         bottom={belowActionBarComponents}
       />
-      {loading ? <IndeterminateLoadingBar /> : null}
+      <IndeterminateLoadingBar loading={loading} />
       {content()}
     </>
   );
